Improved behaviour
Decreased gravity, increased catching food reward x20, 2 new additional states - top and bottom detects if colliding with food. Observed much better training results.
Theorem that stems lack the strength to balance the plant out is supported by decreased gravity. However, it's also possible that the food was moving too quickly for plants to eat it, but as gravity fell, it became incredibly slow, making it much easier for plants to do so.